pyproject.toml
[project]
name = "extractors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.9",
    "beautifulsoup4>=4.13.4",
    "camelot-py[cv]>=1.0.0",
    "lxml>=5.4.0",
    "pandas>=2.3.0",
    "pdfplumber>=0.11.6",
    "playwright>=1.52.0",
    "prompt-toolkit>=3.0.51",
    "pydantic>=2.11.5",
    "pymupdf>=1.26.0",
    "python-anticaptcha>=1.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "selenium>=4.33.0",
    "setuptools>=80.9.0",
]
